--- src/main/java/blackjack/domain/Gamer.java
@@ -1,12 +1,0 @@
-package blackjack.domain;
-
-import blackjack.domain.card.Card;
-
-public interface Gamer {
-
-    void draw(final Card card);
-
-    boolean canDraw();
-
-    int calculateScore();
-}
