+++ src/main/java/blackjack/domain/dto/OutcomeDto.java
@@ -1,6 +1,6 @@
 package blackjack.domain.dto;
 
-import blackjack.domain.Name;
+import blackjack.domain.gamer.Name;
 import blackjack.domain.Outcome;
 
 public class OutcomeDto {
